backend/requirements.txt
asgiref==3.3.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
cryptography==3.2.1
defusedxml==0.6.0
dj-database-url==0.5.0
dj-rest-auth==2.1.2
Django==3.1.4
django-allauth==0.44.0
django-cors-headers==3.6.0
django-debug-toolbar==3.2
django-heroku==0.3.1
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
oauthlib==3.1.0
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
sentry-sdk==0.19.5
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.2.0
